+++ express-zod-api/tests/server-helpers.spec.ts
@@ -21,6 +21,7 @@ import {
   makeResponseMock,
 } from "../src/testing";
 import createHttpError from "http-errors";
+import { AbstractResultHandler } from "../src/result-handler.ts";
 
 describe("Server helpers", () => {
   describe("createCatcher()", () => {
@@ -57,6 +58,40 @@ describe("Server helpers", () => {
         expect(spy.mock.calls[0]![0].error).toEqual(
           error instanceof Error ? error : new Error(error),
         );
+      },
+    );
+
+    test.each([() => fail("I am faulty"), () => Promise.reject("I am faulty")])(
+      "should call Last Resort Handler in case of errorHandler::handler is faulty %#",
+      async (rhImpl) => {
+        const errorHandler = new ResultHandler({
+          positive: vi.fn(),
+          negative: vi.fn(),
+          handler: vi.fn().mockImplementation(rhImpl),
+        });
+        const spy = vi.spyOn(AbstractResultHandler, "lastResort");
+        const handler = createCatcher({
+          errorHandler,
+          getLogger: () => makeLoggerMock(),
+        });
+        const responseMock = makeResponseMock();
+        await handler(
+          new Error("boom"),
+          makeRequestMock(),
+          responseMock,
+          vi.fn(),
+        );
+        expect(spy).toHaveBeenCalledWith({
+          error: expect.objectContaining({
+            message: "I am faulty",
+            cause: expect.objectContaining({
+              message: "I am faulty",
+            }),
+            handled: expect.objectContaining({ message: "boom" }),
+          }),
+          logger: expect.any(Object),
+          response: responseMock,
+        });
       },
     );
   });
@@ -100,7 +135,7 @@ describe("Server helpers", () => {
           negative: vi.fn(),
           handler: vi.fn().mockImplementation(rhImpl),
         });
-        const spy = vi.spyOn(errorHandler, "execute");
+        const spy = vi.spyOn(AbstractResultHandler, "lastResort");
         const handler = createNotFoundHandler({
           errorHandler,
           getLogger: () => makeLoggerMock(),
@@ -114,12 +149,17 @@ describe("Server helpers", () => {
         const responseMock = makeResponseMock();
         await handler(requestMock, responseMock, next);
         expect(next).toHaveBeenCalledTimes(0);
-        expect(spy).toHaveBeenCalledTimes(1);
-        expect(responseMock._getStatusCode()).toBe(500);
-        expect(responseMock._getData()).toBe(
-          "An error occurred while serving the result: I am faulty.\n" +
-            "Original error: Can not POST /v1/test.",
-        );
+        expect(spy).toHaveBeenCalledWith({
+          error: expect.objectContaining({
+            message: "I am faulty",
+            cause: expect.objectContaining({ message: "I am faulty" }),
+            handled: expect.objectContaining({
+              message: "Can not POST /v1/test",
+            }),
+          }),
+          logger: expect.any(Object),
+          response: responseMock,
+        });
       },
     );
   });